Moved env checker into rundev command
